feat: implement search functionality with new controller, service, and utility classes; add OpenNLP and Jackson dependencies

+++ search/src/main/java/com/edward1141/search/entity/ParentChild.java
@@ -12,8 +12,8 @@ import lombok.Data;
 @IdClass(ParentChildId.class)
 public class ParentChild {
     @Id
-    private Integer parentid;
+    private Long parentid;
     
     @Id
-    private Integer childid;
+    private Long childid;
 }
